requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic-core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
cryptography==41.0.7
email-validator==2.1.0
jinja2==3.1.2